+++ src/clj/rems/api/workflows.clj
@@ -10,7 +10,7 @@
 (s/defschema CreateWorkflowCommand
   {:organization s/Str
    :title s/Str
-   :type (s/enum :dynamic)
+   :type (s/enum :workflow/dynamic)
    (s/optional-key :handlers) [UserId]})
 
 (s/defschema EditWorkflowCommand